+++ Assets/Resources/Guns.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1930e4a0880a0b14aab7616a7127fe1e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
