+++ Assets/Resources/Guns/Pistol.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 65461579e77acc141a5eafa3c57cb974
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
