+++ Assets/Resources/Guns/Sniper.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b727d1e46a170db42b95da4a45d09ee7
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
