+++ Assets/Resources/UI/BulletImage.prefab
@@ -12,7 +12,7 @@ GameObject:
   - component: {fileID: 7103305655772931495}
   - component: {fileID: 7103305655772931494}
   m_Layer: 5
-  m_Name: Image
+  m_Name: BulletImage
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
