+++ Assets/Scripts/UI Scripts/AmmoUIScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5e6d2cebc25491b44b7d360a2c449d72
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
